Move `isUserTurn` to its own library
